requirements.txt
pillow
selenium
tqdm